# envoyfilter: clone inputs before patching