… pilot/pkg/xds/testdata/benchmarks/telemetry.yaml