Fixes https://github.com/istio/istio/issues/27826

This avoids the entire class of problems caused by Envoyfilters patching
shared objects. For example, if we define some common XDS component as a
global or shared object, then pass this to multiple different XDS
configs (or even worse, multiple different proxies), the EnvoyFilter
patching will persist in this shared object. This means the patch will
exceed its original scope, both in namespace/workload selector, and
lifetime (removing the EnvoyFilter will do nothing).

In a cursory inspection, I did not actually find any instances of our
code impacted by this, so this is more of a defense in depth. This was
detected in a real environment in older versions of Istio, so its not a
purely theoretical problem.

```
name                             old time/op        new time/op        delta
RouteGeneration/telemetry-6             956µs ± 3%        1285µs ± 3%  +34.40%  (p=0.000 n=10+10)
ClusterGeneration/telemetry-6          5.30ms ± 2%        8.74ms ± 2%  +65.01%  (p=0.000 n=10+10)
ListenerGeneration/telemetry-6         1.63ms ± 3%        1.69ms ± 3%   +4.09%  (p=0.000 n=10+10)
NameTableGeneration/telemetry-6        1.38µs ± 2%        1.37µs ± 3%     ~     (p=0.101 n=10+10)

name                             old alloc/op       new alloc/op       delta
RouteGeneration/telemetry-6             619kB ± 0%         800kB ± 0%  +29.15%  (p=0.000 n=10+9)
ClusterGeneration/telemetry-6          2.24MB ± 0%        3.53MB ± 0%  +57.58%  (p=0.000 n=9+8)
ListenerGeneration/telemetry-6          824kB ± 0%         855kB ± 0%   +3.73%  (p=0.000 n=10+9)
NameTableGeneration/telemetry-6          265B ± 0%          265B ± 0%     ~     (all equal)
```

This comes with a pretty high performance cost; we may need to re-evaluate ways to reduce the impact
such as shallow copying where possible, skipping cloning when there
are no patches to apply, etc



[ ] Configuration Infrastructure
[ ] Docs
[ ] Installation
[ ] Networking
[ ] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[ ] User Experience
[ ] Developer Infrastructure


Pull Request Attributes

Please check any characteristics that apply to this pull request. 

[ ] Does not have any changes that may affect Istio users.